refactor: move RocksDBWrapper to broker
RocksDBWrapper is used in Broker tests only.
