+++ broker/src/test/java/io/camunda/zeebe/broker/system/partitions/impl/RocksDBWrapper.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.logstreams.util;
+package io.camunda.zeebe.broker.system.partitions.impl;
 
 import io.camunda.zeebe.db.ColumnFamily;
 import io.camunda.zeebe.db.ZeebeDb;
